Move tree properties to tree_properties.jl
